+++ extensions/github/src/components/RepositoryActions.tsx
@@ -5,13 +5,13 @@ import { format } from "date-fns";
 import { getGitHubClient } from "../api/githubClient";
 import { ExtendedRepositoryFieldsFragment } from "../generated/graphql";
 import { getErrorMessage } from "../helpers/errors";
-import { cloneAndOpen, REPO_SORT_TYPES_TO_QUERIES, WEB_IDES } from "../helpers/repository";
+import { cloneAndOpen, WEB_IDES } from "../helpers/repository";
 
 import { RepositoryDiscussionList } from "./RepositoryDiscussions";
 import { RepositoryIssueList } from "./RepositoryIssues";
 import { RepositoryPullRequestList } from "./RepositoryPullRequest";
 import RepositoryReleases from "./RepositoryReleases";
-import { SortAction, SortActionProps } from "./SortAction";
+import { SortAction, SortActionProps, SortTypeProps } from "./SortAction";
 
 type RepositoryActionProps = {
   repository: ExtendedRepositoryFieldsFragment;
@@ -25,7 +25,8 @@ export default function RepositoryActions({
   onVisit,
   setSortQuery,
   sortQuery,
-}: RepositoryActionProps & SortActionProps) {
+  sortTypes,
+}: RepositoryActionProps & SortActionProps & SortTypeProps) {
   const { github } = getGitHubClient();
 
   const updatedAt = new Date(repository.updatedAt);
@@ -239,7 +240,7 @@ export default function RepositoryActions({
       </ActionPanel.Section>
 
       <ActionPanel.Section>
-        <SortAction data={REPO_SORT_TYPES_TO_QUERIES} {...{ sortQuery, setSortQuery }} />
+        <SortAction {...{ data: sortTypes, sortQuery, setSortQuery }} />
       </ActionPanel.Section>
     </ActionPanel>
   );