+++ extensions/github/CHANGELOG.md
@@ -1,8 +1,9 @@
 # GitHub Changelog
 
-## [Repository Search] - {PR_MERGE_DATE}
+## [Repository Improvements] - {PR_MERGE_DATE}
 
 - Adds primary language color and code icon to the accessory for being in-sync with GitHub UI.
+- Adds "Sort By" action for My Latest Repositories with supported options
 
 ## [Improvements and Fixes] - 2024-07-08